+++ src/htscoremain.c
@@ -45,9 +45,7 @@ Please visit our Website: http://www.httrack.com
 #include "htsmodules.h"
 #include "htszlib.h"
 #include "htscharset.h"
-#include "htsencoding.h"
-#include "htscache_selftest.h"
-#include "htsdns_selftest.h"
+#include "htsselftest.h"
 #include "htsmd5.h"
 
 #include <ctype.h>
@@ -112,450 +110,6 @@ HTSEXT_API int hts_main(int argc, char **argv) {
 
   hts_free_opt(opt);
   return ret;
-}
-
-// very minimalistic internal tests
-static void basic_selftests(void) {
-  // BUG 756328
-  const char *const source = "/intent/tweet?url=https%3A%2F%2Fwww.httrack.com%2Fvacatures%2F1562519%2Fmedewerker-data-services&text=Medewerker+Data+Services&via=httrackcom";
-  char buffer[1024];
-  fil_normalized(source, buffer);
-  // MD5 selftests
-  md5selftest();
-  // cookie_get field extraction (tab-separated, 0-based)
-  {
-    char cbuf[8192];
-
-    assertf(strcmp(cookie_get(cbuf, "a\tb\tc", 0), "a") == 0);
-    assertf(strcmp(cookie_get(cbuf, "a\tb\tc", 1), "b") == 0);
-    assertf(strcmp(cookie_get(cbuf, "a\tb\tc", 2), "c") == 0);
-    // multi-char fields catch length/boundary bugs that 1-char fields hide
-    assertf(strcmp(cookie_get(cbuf, "host\tx\t/path/to", 0), "host") == 0);
-    assertf(strcmp(cookie_get(cbuf, "host\tx\t/path/to", 2), "/path/to") == 0);
-    assertf(strcmp(cookie_get(cbuf, "a\t\tc", 1), "") == 0);  // empty field
-    assertf(strcmp(cookie_get(cbuf, "a\tb\tc", 9), "") == 0); // beyond last
-  }
-  // back_infostr() status-line formatting (no sockets: pure formatting over
-  // in-memory slots). Stresses a few thousand entries across every status-code
-  // arm. Regression for a clobber bug where the size/totalsize trailer was
-  // written straight into the destination, wiping the URL it had just built.
-  {
-    static const struct {
-      int code;
-      const char *tag;
-    } cases[] = {
-        {200, "READY "},     {-1, "ERROR "},       {-2, "TIMEOUT "},
-        {-3, "TOOSLOW "},    {400, "BADREQUEST "}, {403, "FORBIDDEN "},
-        {404, "NOT FOUND "}, {500, "SERVERROR "},  {999, "ERROR(999)"},
-    };
-    const int ncases = (int) (sizeof(cases) / sizeof(cases[0]));
-    const int n = 2000;
-    lien_back *slots = calloct(n, sizeof(lien_back));
-    char line[HTS_URLMAXSIZE * 4 + 1024];
-    char expect[HTS_URLMAXSIZE * 4 + 1024];
-    struct_back sb;
-    int idx;
-
-    sb.lnk = slots;
-    sb.count = n;
-    sb.ready = NULL;
-    sb.ready_size_bytes = 0;
-    for (idx = 0; idx < n; idx++) {
-      lien_back *const slot = &slots[idx];
-
-      slot->r.location = slot->location_buffer;
-      slot->status = STATUS_READY;
-      slot->r.statuscode = cases[idx % ncases].code;
-      slot->r.size = idx;
-      slot->r.totalsize = idx + 1;
-      snprintf(slot->url_adr, sizeof(slot->url_adr), "http://h%d.example", idx);
-      snprintf(slot->url_fil, sizeof(slot->url_fil), "/p/%d.html", idx);
-    }
-    for (idx = 0; idx < n; idx++) {
-      line[0] = '\0';
-      back_infostr(&sb, idx, 3, line, sizeof(line));
-      // Exact match (not substring): pins tag/URL/trailer order and rejects a
-      // partial clobber, duplication, or truncation that a presence check would
-      // let through. The expected format is stated here independently.
-      snprintf(expect, sizeof(expect),
-               "%s\"http://h%d.example/p/%d.html\" " LLintP " " LLintP " ",
-               cases[idx % ncases].tag, idx, idx, (LLint) idx,
-               (LLint) (idx + 1));
-      assertf(strcmp(line, expect) == 0);
-    }
-    // Near-maximal URL, driven through back_info() (which owns the status
-    // buffer internally and prints to a FILE*). url_adr + url_fil together
-    // overrun the old HTS_URLMAXSIZE*2+1024 buffer, so the bounded appends
-    // would abort unless that buffer is sized to hold both fields. Regression
-    // for that sizing -- exercising back_infostr() directly would miss it,
-    // since the caller's buffer is what matters.
-    {
-      lien_back *const slot = &slots[0];
-      const size_t adrlen = sizeof(slot->url_adr) - 8;
-      const size_t fillen = sizeof(slot->url_fil) - 8;
-      FILE *const fp = tmpfile();
-      size_t got;
-
-      assertf(fp != NULL);
-      slot->status = STATUS_READY;
-      slot->r.statuscode = 200;
-      slot->r.size = 1;
-      slot->r.totalsize = 2;
-      memset(slot->url_adr, 'a', adrlen);
-      slot->url_adr[adrlen] = '\0';
-      slot->url_fil[0] = '/';
-      memset(slot->url_fil + 1, 'b', fillen - 1);
-      slot->url_fil[fillen] = '\0';
-      back_info(&sb, 0, 3, fp);
-      rewind(fp);
-      got = fread(line, 1, sizeof(line) - 1, fp);
-      line[got] = '\0';
-      fclose(fp);
-      snprintf(expect, sizeof(expect),
-               "READY \"%s%s\" " LLintP " " LLintP " " LF, slot->url_adr,
-               slot->url_fil, (LLint) 1, (LLint) 2);
-      assertf(strcmp(line, expect) == 0);
-    }
-    freet(slots);
-  }
-  // next_token(): in-place token scanner. Strips surrounding quotes, unescapes
-  // \" and \\ when flag is set, and returns the token terminator (the space, or
-  // NULL at end of string). The unquote/unescape rewrites the string in place
-  // by shifting left, so the result is always shorter -- regression for that
-  // compaction.
-  {
-    char tok[64];
-
-    // plain token: unchanged, returns a pointer AT the separating space (exact
-    // position, not just any space -- a strchr-style impl would land elsewhere
-    // once quotes shift the content)
-    strcpybuff(tok, "abc def");
-    {
-      char *const end = next_token(tok, 0);
-      assertf(end == tok + 3 && *end == ' ' && strcmp(tok, "abc def") == 0);
-    }
-    // surrounding quotes stripped, returns the (post-shift) trailing space
-    strcpybuff(tok, "\"ab\" cd");
-    {
-      char *const end = next_token(tok, 1);
-      assertf(end == tok + 2 && *end == ' ' && strcmp(tok, "ab cd") == 0);
-    }
-    // a space inside quotes does not end the token; end of string returns NULL
-    strcpybuff(tok, "\"a b\"c");
-    {
-      char *const end = next_token(tok, 1);
-      assertf(end == NULL && strcmp(tok, "a bc") == 0);
-    }
-    // \" and \\ are unescaped to literal " and \ in place
-    strcpybuff(tok, "\"a\\\"b\\\\c\"");
-    {
-      char *const end = next_token(tok, 1);
-      assertf(end == NULL && strcmp(tok, "a\"b\\c") == 0);
-    }
-    // unterminated quote: the opening quote is dropped, the rest survives, and
-    // the scan runs to the NUL (returns NULL)
-    strcpybuff(tok, "\"ab");
-    {
-      char *const end = next_token(tok, 1);
-      assertf(end == NULL && strcmp(tok, "ab") == 0);
-    }
-    // trailing lone backslash in a quote: *(p+1) is the NUL, not an escape, so
-    // the backslash is kept intact (and there is no over-read past the NUL)
-    strcpybuff(tok, "\"a\\");
-    {
-      char *const end = next_token(tok, 1);
-      assertf(end == NULL && strcmp(tok, "a\\") == 0);
-    }
-  }
-  // fil_normalized(): canonicalizes a URL path. Query arguments are sorted
-  // alphabetically (by the text after each '?'/'&') and the query is rebuilt
-  // through a bounded builder; outside the query, "//" collapses to "/".
-  // Regression for that builder.
-  {
-    char norm[256];
-
-    assertf(strcmp(fil_normalized("/p?b=2&a=1&c=3", norm), "/p?a=1&b=2&c=3") ==
-            0);
-    assertf(strcmp(fil_normalized("/a//b", norm), "/a/b") == 0);
-    // "//" is collapsed only before the query; inside the query it is kept
-    assertf(strcmp(fil_normalized("/a//b?x=c//d", norm), "/a/b?x=c//d") == 0);
-  }
-  // give_mimext(): mime type -> file extension, bounded into the caller buffer.
-  // Returns 1 when an extension was written, 0 otherwise.
-  {
-    char ext[16];
-
-    assertf(give_mimext(ext, sizeof(ext), "image/gif") == 1);
-    assertf(strcmp(ext, "gif") == 0);
-    assertf(give_mimext(ext, sizeof(ext), "text/html") == 1);
-    assertf(strcmp(ext, "html") == 0);
-    assertf(give_mimext(ext, sizeof(ext), "no/such-mime-type") == 0);
-    assertf(ext[0] == '\0');
-  }
-  // convtolower(): lower-cases into the caller buffer (bounded by its size).
-  {
-    char low[64];
-
-    assertf(strcmp(convtolower(low, sizeof(low), "ABC/Def.HTML"),
-                   "abc/def.html") == 0);
-  }
-  // cut_path(): splits a path into directory (with trailing '/') and basename,
-  // each bounded by its buffer size.
-  {
-    char path[256];
-    char pname[256];
-
-    {
-      char full[] = "/dir/sub/file.html";
-
-      cut_path(full, path, sizeof(path), pname, sizeof(pname));
-      assertf(strcmp(path, "/dir/sub/") == 0);
-      assertf(strcmp(pname, "file.html") == 0);
-    }
-    { // a trailing slash is trimmed before the split
-      char full[] = "/dir/sub/";
-
-      cut_path(full, path, sizeof(path), pname, sizeof(pname));
-      assertf(strcmp(path, "/dir/") == 0);
-      assertf(strcmp(pname, "sub") == 0);
-    }
-    { // a path of length <= 1 yields empty results
-      char full[] = "/";
-
-      cut_path(full, path, sizeof(path), pname, sizeof(pname));
-      assertf(path[0] == '\0' && pname[0] == '\0');
-    }
-  }
-  // get_httptype_sized(): a long MIME type (Office OOXML reaches 73 chars) is
-  // written whole into a contenttype-sized buffer; returns 1 on a match, 0 when
-  // flag==0 and nothing matched. Regression for the old contenttype[64]
-  // overflow.
-  {
-    httrackp *opt = hts_create_opt();
-    htsblk r; // write into the real struct field, not a stand-in
-
-    assertf(opt != NULL);
-    // a long MIME (Office OOXML reaches 73 chars) must fit htsblk.contenttype
-    // whole: a [64] field would make this bounded copy abort.
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "deck.pptx", 0) == 1);
-    assertf(strcmp(r.contenttype,
-                   "application/vnd.openxmlformats-officedocument."
-                   "presentationml.presentation") == 0);
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "x.gif", 0) == 1);
-    assertf(strcmp(r.contenttype, "image/gif") == 0);
-    // no extension and flag==0: nothing written, returns 0
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "noextfile", 0) == 0);
-    assertf(r.contenttype[0] == '\0');
-    // no extension and flag==1: octet-stream fallback, returns 1
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "noextfile", 1) == 1);
-    assertf(strcmp(r.contenttype, "application/octet-stream") == 0);
-    // empty fil: no extension to scan; must not over-read before the string.
-    // flag==0 -> 0 (nothing written), flag==1 -> octet-stream.
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype), "",
-                               0) == 0);
-    assertf(r.contenttype[0] == '\0');
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype), "",
-                               1) == 1);
-    assertf(strcmp(r.contenttype, "application/octet-stream") == 0);
-    // a user --assume rule with an empty value matches but writes nothing:
-    // get_userhttptype returns 1 with the buffer empty, so get_httptype_sized
-    // must still report 0 (callers test the return like the old
-    // strnotempty(s)).
-    StringCopy(opt->mimedefs, "\ncgi=\n");
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "/x.cgi", 0) == 0);
-    assertf(r.contenttype[0] == '\0');
-    StringCopy(opt->mimedefs, "\ncgi=text/html\n");
-    assertf(get_httptype_sized(opt, r.contenttype, sizeof(r.contenttype),
-                               "/x.cgi", 0) == 1);
-    assertf(strcmp(r.contenttype, "text/html") == 0);
-    hts_free_opt(opt);
-  }
-  // adr_normalized_sized(): bounded host normalization (passthrough when
-  // already normal).
-  {
-    char n[HTS_URLMAXSIZE];
-
-    assertf(strcmp(adr_normalized_sized("example.com", n, sizeof(n)),
-                   "example.com") == 0);
-  }
-  // standard_name(): builds "<name><md5?>.<ext>" into a bounded buffer. The md5
-  // is appended (4 chars) only when the URL has a query string (see url_md5),
-  // so test both; pin the structure (name + ext, lengths), not the md5 chars.
-  {
-    char b[HTS_URLMAXSIZE * 2];
-    const char *nom = "index.html"; // name part
-    const char *dot = nom + 5;      // points at ".html"
-    size_t len;
-
-    // no query -> no md5: "index" + ".html"
-    standard_name(b, sizeof(b), dot, nom, "http://example.com/index.html", 0);
-    assertf(strcmp(b, "index.html") == 0);
-    // query -> 4 md5 chars between name and ext: "index" + md5(4) + ".html"
-    standard_name(b, sizeof(b), dot, nom, "http://example.com/index.html?v=1",
-                  0);
-    len = strlen(b);
-    assertf(len == 5 + 4 + 5);
-    assertf(strncmp(b, "index", 5) == 0);
-    assertf(strcmp(b + len - 5, ".html") == 0);
-    // short names: name kept (<=8), the extension is clamped to 3 -> ".htm"
-    standard_name(b, sizeof(b), dot, nom, "http://example.com/index.html?v=1",
-                  1);
-    len = strlen(b);
-    assertf(len == 5 + 4 + 4);
-    assertf(strcmp(b + len - 4, ".htm") == 0);
-    // short names with a >8-char name: the name is clamped to 8 ("indexpag")
-    {
-      const char *lnom = "indexpage.html";
-      const char *ldot = lnom + 9; // points at ".html"
-
-      standard_name(b, sizeof(b), ldot, lnom,
-                    "http://example.com/indexpage.html?v=1", 1);
-      len = strlen(b);
-      assertf(len == 8 + 4 + 4);
-      assertf(strncmp(b, "indexpag", 8) == 0);
-      assertf(strcmp(b + len - 4, ".htm") == 0);
-    }
-  }
-  // longfile_to_83(): single-name 8-3 (mode 1) / ISO9660 (mode 2) conversion;
-  // uppercases, clamps the name (8 / 31) and the extension (3). It rewrites
-  // 'save' in place, so pass a mutable array.
-  {
-    char n83[256];
-
-    {
-      char save[] = "longfilename.html";
-
-      longfile_to_83(1, n83, sizeof(n83), save); // 8-3: name->8, ext->3
-      assertf(strcmp(n83, "LONGFILE.HTM") == 0);
-    }
-    {
-      char save[] = "longfilename.html";
-
-      longfile_to_83(2, n83, sizeof(n83), save); // ISO9660: name->31, ext->3
-      assertf(strcmp(n83, "LONGFILENAME.HTM") == 0);
-    }
-    { // sanitization: leading '.'->'_', interior dots
-      char save[] = ".a b.c.d e"; // collapse to '_', spaces/specials -> '_'
-                                  // (only the last dot stays as the separator)
-      longfile_to_83(1, n83, sizeof(n83), save);
-      assertf(strcmp(n83, "_A_B_C.D_E") == 0);
-    }
-  }
-  // long_to_83(): per-segment 8-3 conversion of a whole path.
-  {
-    char n83[HTS_URLMAXSIZE * 2];
-    char save[] = "dir/longfilename.html";
-
-    long_to_83(1, n83, sizeof(n83), save);
-    assertf(strcmp(n83, "DIR/LONGFILE.HTM") == 0);
-  }
-  // lienrelatif(): relative path from the directory of curr_fil to link.
-  {
-    char s[HTS_URLMAXSIZE * 2];
-
-    // same directory -> just the basename
-    assertf(lienrelatif(s, sizeof(s), "dir/page.html", "dir/index.html") == 0);
-    assertf(strcmp(s, "page.html") == 0);
-    // link one level up -> a "../" prefix
-    assertf(lienrelatif(s, sizeof(s), "a.html", "dir/index.html") == 0);
-    assertf(strcmp(s, "../a.html") == 0);
-  }
-}
-
-/* Self-tests for the htssafe.h bounded string ops (driven by httrack -#8).
-   Returns 0 if every bounded operation behaved correctly, 1 otherwise.
-   The abort-on-overflow guarantee is checked separately by the -#8 "overflow"
-   sub-mode (it aborts the process by design). */
-static int string_safety_selftests(void) {
-  char buf[8];
-
-  /* strcpybuff into a sized array: exact copy */
-  strcpybuff(buf, "abc");
-  if (strcmp(buf, "abc") != 0)
-    return 1;
-
-  /* strcatbuff append within capacity */
-  strcatbuff(buf, "de");
-  if (strcmp(buf, "abcde") != 0)
-    return 1;
-
-  /* strncatbuff appends at most N source chars */
-  strcpybuff(buf, "ab");
-  strncatbuff(buf, "cdef", 2);
-  if (strcmp(buf, "abcd") != 0)
-    return 1;
-
-  /* strlcpybuff: explicit-capacity copy into a pointer destination, the form
-     the migration moves toward */
-  {
-    char storage[8];
-    char *const p = storage;
-
-    strlcpybuff(p, "hello", sizeof(storage));
-    if (strcmp(p, "hello") != 0)
-      return 1;
-  }
-
-  /* strcpybuff into a pointer destination: routes through the unchecked
-     strcpybuff_ptr_ fallback (the path the -#8 warning flags). The warning is
-     intentional here; we only verify the fallback still copies correctly. */
-#if defined(__GNUC__)
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wattribute-warning"
-#endif
-  {
-    char storage[8];
-    char *const p = storage;
-
-    strcpybuff(p, "ptr");
-    if (strcmp(p, "ptr") != 0)
-      return 1;
-  }
-#if defined(__GNUC__)
-#pragma GCC diagnostic pop
-#endif
-
-  /* htsbuff: bounded builder over a fixed array (append, truncating append,
-     reset, and length tracking) */
-  {
-    char dst[8];
-    htsbuff b = htsbuff_array(dst);
-
-    htsbuff_cat(&b, "ab");
-    htsbuff_cat(&b, "cd");
-    if (strcmp(htsbuff_str(&b), "abcd") != 0 || b.len != 4)
-      return 1;
-
-    htsbuff_catn(&b, "efghij", 2);      /* append at most 2 */
-    if (strcmp(htsbuff_str(&b), "abcdef") != 0)
-      return 1;
-
-    htsbuff_cpy(&b, "xyz");             /* reset */
-    if (strcmp(htsbuff_str(&b), "xyz") != 0 || b.len != 3)
-      return 1;
-
-    htsbuff_catc(&b, '!'); /* single character */
-    if (strcmp(htsbuff_str(&b), "xyz!") != 0 || b.len != 4)
-      return 1;
-  }
-
-  /* boundary: filling to exactly cap-1 must succeed (one more aborts, which the
-     -#8 overflow-buff mode checks) */
-  {
-    char d2[4];
-    htsbuff c = htsbuff_array(d2);
-
-    htsbuff_cat(&c, "abc");
-    if (strcmp(htsbuff_str(&c), "abc") != 0 || c.len != 3)
-      return 1;
-  }
-
-  return 0;
 }
 
 static int hts_main_internal(int argc, char **argv, httrackp * opt);
@@ -1351,6 +905,25 @@ static int hts_main_internal(int argc, char **argv, httrackp * opt) {
      }
      }
    */
+
+  /* Engine self-tests: -#test lists them, -#test=NAME [args] runs one. Handled
+     here, ahead of the no-URL usage gate below, so they need no dummy URL. */
+  {
+    int k;
+
+    for (k = 1; k < argc; k++) {
+      const char *const a = argv[k];
+
+      if (a[0] == '-' && a[1] == '#' && strncmp(a + 2, "test", 4) == 0 &&
+          (a[6] == '\0' || a[6] == '=')) {
+        const char *const name = a[6] == '=' ? a + 7 : NULL;
+        const int code = hts_selftest(opt, name, argc - (k + 1), &argv[k + 1]);
+
+        htsmain_free();
+        return code;
+      }
+    }
+  }
 
   // Pas d'URL
 #if DEBUG_STEPS
@@ -2440,94 +2013,6 @@ static int hts_main_internal(int argc, char **argv, httrackp * opt) {
           case '#':{           // non documenté
               com++;
               switch (*com) {
-              case 'A': // cache self-test: httrack -#A <dir>
-                if (na + 1 < argc) {
-                  const int err = cache_selftests(opt, argv[na + 1]);
-
-                  printf("cache-selftest: %s\n", err ? "FAIL" : "OK");
-                  htsmain_free();
-                  return err;
-                } else {
-                  fprintf(stderr, "Option #A requires a directory argument\n");
-                  htsmain_free();
-                  return 1;
-                }
-                break;
-              case 'W': // cache write-failure handling: httrack -#W <dir>
-                if (na + 1 < argc) {
-                  const int err =
-                      cache_write_failure_selftest(opt, argv[na + 1]);
-
-                  printf("cache-writefail: %s\n", err ? "FAIL" : "OK");
-                  htsmain_free();
-                  return err;
-                } else {
-                  fprintf(stderr, "Option #W requires a directory argument\n");
-                  htsmain_free();
-                  return 1;
-                }
-                break;
-              case 'B': // golden cache fixture read: httrack -#B <dir> [regen]
-                if (na + 1 < argc) {
-                  const int regen =
-                      (na + 2 < argc && strcmp(argv[na + 2], "regen") == 0);
-                  const int err =
-                      cache_golden_selftest(opt, argv[na + 1], regen);
-
-                  printf("cache-golden: %s\n", err ? "FAIL" : "OK");
-                  htsmain_free();
-                  return err;
-                } else {
-                  fprintf(stderr, "Option #B requires a directory argument\n");
-                  htsmain_free();
-                  return 1;
-                }
-                break;
-              case 'D': { // DNS resolver/cache self-test (mock getaddrinfo)
-                const int err = dns_selftests(opt);
-
-                printf("dns-selftest: %s\n", err ? "FAIL" : "OK");
-                htsmain_free();
-                return err;
-              } break;
-              case 'N': { // url_savename name resolution: httrack -#N <fil>
-                          // <content-type>
-                if (na + 2 < argc) {
-                  lien_adrfilsave afs;
-                  cache_back cache;
-                  struct_back *sback;
-                  hash_struct hash;
-                  lien_back headers;
-
-                  memset(&afs, 0, sizeof(afs));
-                  strcpybuff(afs.af.adr, "www.example.com");
-                  strcpybuff(afs.af.fil, argv[na + 1]);
-
-                  memset(&cache, 0, sizeof(cache));
-                  cache.hashtable = (void *) coucal_new(0);
-
-                  sback = back_new(opt, opt->maxsoc * 32 + 1024);
-                  hash_init(opt, &hash, opt->urlhack);
-
-                  memset(&headers, 0, sizeof(headers));
-                  headers.status = 0;
-                  headers.r.statuscode = HTTP_OK;
-                  strcpybuff(headers.r.contenttype, argv[na + 2]);
-                  strcpybuff(headers.url_fil, argv[na + 1]);
-
-                  url_savename(&afs, NULL, NULL, NULL, opt, sback, &cache,
-                               &hash, 0, 0, &headers);
-                  printf("savename: %s\n", afs.save);
-                  htsmain_free();
-                  return 0;
-                } else {
-                  fprintf(
-                      stderr,
-                      "Option #N requires <fil> <content-type> arguments\n");
-                  htsmain_free();
-                  return 1;
-                }
-              } break;
               case 'C':        // list cache files : httrack -#C '*spid*.gif' will attempt to find the matching file
                 {
                   int hasFilter = 0;
@@ -2825,468 +2310,6 @@ static int hts_main_internal(int argc, char **argv, httrackp * opt) {
                 }
                 break;
 
-              case '0':        /* test #0 : filters */
-                if (na + 2 >= argc) {
-                  HTS_PANIC_PRINTF
-                    ("Option #0 needs to be followed by a filter string and a string");
-                  printf("Example: '-#0' '*.gif' 'foo.gif'\n");
-                  htsmain_free();
-                  return -1;
-                } else {
-                  if (strjoker(argv[na + 2], argv[na + 1], NULL, NULL))
-                    printf("%s does match %s\n", argv[na + 2], argv[na + 1]);
-                  else
-                    printf("%s does NOT match %s\n", argv[na + 2],
-                           argv[na + 1]);
-                  htsmain_free();
-                  return 0;
-                }
-                break;
-              case '1':        /* test #1 : fil_simplifie */
-                if (na + 1 >= argc) {
-                  HTS_PANIC_PRINTF("Option #1 needs to be followed by an URL");
-                  printf("Example: '-#1' ./foo/bar/../foobar\n");
-                  htsmain_free();
-                  return -1;
-                } else {
-                  fil_simplifie(argv[na + 1]);
-                  printf("simplified=%s\n", argv[na + 1]);
-                  htsmain_free();
-                  return 0;
-                }
-                break;
-              case 'l': /* lienrelatif: relative link from curr_fil to link */
-                if (na + 2 >= argc) {
-                  HTS_PANIC_PRINTF(
-                      "Option #l needs a link and a current-file path");
-                  printf(
-                      "Example: '-#l' 'host/dir/img.gif' 'host/dir/p.html'\n");
-                  htsmain_free();
-                  return -1;
-                } else {
-                  char s[HTS_URLMAXSIZE * 2];
-
-                  if (lienrelatif(s, sizeof(s), argv[na + 1], argv[na + 2]) ==
-                      0)
-                    printf("relative=%s\n", s);
-                  else
-                    printf("relative=<ERROR>\n");
-                  htsmain_free();
-                  return 0;
-                }
-                break;
-              case 'i': /* ident_url_relatif: resolve a link -> adr/fil */
-                if (na + 3 >= argc) {
-                  HTS_PANIC_PRINTF(
-                      "Option #i needs a link, an origin address and file");
-                  printf("Example: '-#i' '../img.gif' 'www.foo.com' "
-                         "'/d/p.html'\n");
-                  htsmain_free();
-                  return -1;
-                } else {
-                  lien_adrfil af;
-                  const int r = ident_url_relatif(argv[na + 1], argv[na + 2],
-                                                  argv[na + 3], &af);
-
-                  if (r == 0)
-                    printf("adr=%s fil=%s\n", af.adr, af.fil);
-                  else
-                    printf("error=%d\n", r);
-                  htsmain_free();
-                  return 0;
-                }
-                break;
-              case '2':        // mimedefs
-                if (na + 1 >= argc) {
-                  HTS_PANIC_PRINTF("Option #2 needs to be followed by an URL");
-                  printf("Example: '-#2' /foo/bar.php\n");
-                  htsmain_free();
-                  return -1;
-                } else {
-                  char mime[256];
-
-                  // initialiser mimedefs
-                  //get_userhttptype(opt,1,opt->mimedefs,NULL);
-                  // check
-                  if (get_httptype_sized(opt, mime, sizeof(mime), argv[na + 1],
-                                         0)) {
-                    char ext[256];
-
-                    printf("%s is '%s'\n", argv[na + 1], mime);
-                    if (give_mimext(ext, sizeof(ext), mime)) {
-                      printf("and its local type is '.%s'\n", ext);
-                    }
-                  } else {
-                    printf("%s is of an unknown MIME type\n", argv[na + 1]);
-                  }
-                  htsmain_free();
-                  return 0;
-                }
-                break;
-              case '3':        // charset tests: httrack -#3 "iso-8859-1" "café"
-                if (++na + 1 < argc) {
-                  char *s =
-                    hts_convertStringToUTF8(argv[na+1], strlen(argv[na+1]), argv[na]);
-                  if (s != NULL) {
-                    printf("%s\n", s);
-                    free(s);
-                  } else {
-                    fprintf(stderr, "invalid string for charset %s\n", argv[na]);
-                  }
-                  na += 2;
-                } else {
-                  fprintf(stderr,
-                    "Option #3 needs to be followed by a charset and a string");
-                }
-                htsmain_free();
-                return 0;
-                break;
-              case '4':  // IDNA encoder: httrack -#4 "www.café.com"
-                if (++na < argc) {
-                  char *s = hts_convertStringUTF8ToIDNA(argv[na], strlen(argv[na]));
-                  if (s != NULL) {
-                    printf("%s\n", s);
-                    free(s);
-                  } else {
-                    fprintf(stderr, "invalid string '%s'\n", argv[na]);
-                  }
-                  na += 1;
-                } else {
-                  fprintf(stderr,
-                    "Option #4 needs to be followed by an IDNA string");
-                }
-                htsmain_free();
-                return 0;
-                break;
-              case '5':  // IDNA encoder: httrack -#5
-                if (++na < argc) {
-                  char *s = hts_convertStringIDNAToUTF8(argv[na], strlen(argv[na]));
-                  if (s != NULL) {
-                    printf("%s\n", s);
-                    free(s);
-                  } else {
-                    fprintf(stderr, "invalid string '%s'\n", argv[na]);
-                  }
-                  na += 1;
-                } else {
-                  fprintf(stderr,
-                    "Option #5 needs to be followed by an IDNA string");
-                }
-                htsmain_free();
-                return 0;
-                break;
-              case '6':  // entities: httrack -#6 "&foo;" ["encoding"]
-                if (++na < argc) {
-                  char *const s = strdup(argv[na]);
-                  const char *const enc = na + 1 < argc ? argv[na + 1] : "UTF-8";
-                  if (s != NULL 
-                    && hts_unescapeEntitiesWithCharset(s, s, strlen(s), 
-                                                       enc) == 0) {
-                    printf("%s\n", s);
-                    free(s);
-                  } else {
-                    fprintf(stderr, "invalid string '%s'\n", argv[na]);
-                  }
-                  na += 1;
-                } else {
-                  fprintf(stderr,
-                    "Option #6 needs to be followed by a string");
-                }
-                htsmain_free();
-                return 0;
-                break;
-              case '8':        /* string-safety selftest: httrack -#8 [overflow <bigstr>] */
-                if (na + 1 < argc
-                    && strncmp(argv[na + 1], "overflow", 8) == 0) {
-                  /* Deliberately exceed a sized buffer: the bounded op must
-                     abort. The source comes from argv so its length is opaque
-                     to the compiler (no static -Wstringop-overflow, genuine
-                     runtime check). "overflow-buff" exercises htsbuff. */
-                  char small[4];
-                  const char *const src =
-                    (na + 2 < argc) ? argv[na + 2] : "overflowing";
-
-                  if (strcmp(argv[na + 1], "overflow-buff") == 0) {
-                    htsbuff b = htsbuff_array(small);
-
-                    htsbuff_cat(&b, src);
-                  } else {
-                    strcpybuff(small, src);
-                  }
-                  printf("strsafe: NOT aborted\n");     /* must be unreachable */
-                  htsmain_free();
-                  return 1;
-                } else {
-                  const int err = string_safety_selftests();
-
-                  printf("strsafe: %s\n", err ? "FAIL" : "OK");
-                  htsmain_free();
-                  return err;
-                }
-                break;
-              case '7':  // hashtable selftest: httrack -#7 nb_entries
-                basic_selftests();
-                if (++na < argc) {
-                  char *const snum = strdup(argv[na]);
-                  unsigned long count = 0;
-                  const char *const names[] = {
-                    "", "add", "delete", "dry-add", "dry-del",
-                    "test-exists", "test-not-exist"
-                  };
-                  const struct {
-                    enum {
-                      DO_END,
-                      DO_ADD,
-                      DO_DEL,
-                      DO_DRY_ADD,
-                      DO_DRY_DEL,
-                      TEST_ADD,
-                      TEST_DEL
-                    } type;
-                    size_t modulus;
-                    size_t offset;
-                  } bench[] = {
-                    { DO_ADD, 4, 0 },     /* add 4/0 */
-                    { TEST_ADD, 4, 0 },   /* check 4/0 */
-                    { TEST_DEL, 4, 1 },   /* check 4/1 */
-                    { TEST_DEL, 4, 2 },   /* check 4/2 */
-                    { TEST_DEL, 4, 3 },   /* check 4/3 */
-                    { DO_DRY_DEL, 4, 1 }, /* del 4/1 */
-                    { DO_DRY_DEL, 4, 2 }, /* del 4/2 */
-                    { DO_DRY_DEL, 4, 3 }, /* del 4/3 */
-                    { DO_ADD, 4, 1 },     /* add 4/1 */
-                    { DO_DRY_ADD, 4, 1 }, /* add 4/1 */
-                    { TEST_ADD, 4, 0 },   /* check 4/0 */
-                    { TEST_ADD, 4, 1 },   /* check 4/1 */
-                    { TEST_DEL, 4, 2 },   /* check 4/2 */
-                    { TEST_DEL, 4, 3 },   /* check 4/3 */
-                    { DO_ADD, 4, 2 },     /* add 4/2 */
-                    { DO_DRY_DEL, 4, 3 }, /* del 4/3 */
-                    { DO_ADD, 4, 3 },     /* add 4/3 */
-                    { DO_DEL, 4, 3 },     /* del 4/3 */
-                    { TEST_ADD, 4, 0 },   /* check 4/0 */
-                    { TEST_ADD, 4, 1 },   /* check 4/1 */
-                    { TEST_ADD, 4, 2 },   /* check 4/2 */
-                    { TEST_DEL, 4, 3 },   /* check 4/3 */
-                    { DO_DEL, 4, 0 },     /* del 4/0 */
-                    { DO_DEL, 4, 1 },     /* del 4/1 */
-                    { DO_DEL, 4, 2 },     /* del 4/2 */
-                    /* empty here */
-                    { TEST_DEL, 1, 0 },   /* check */
-                    { DO_ADD, 4, 0 },     /* add 4/0 */
-                    { DO_ADD, 4, 1 },     /* add 4/1 */
-                    { DO_ADD, 4, 2 },     /* add 4/2 */
-                    { DO_DEL, 42, 0 },    /* add 42/0 */
-                    { TEST_DEL, 42, 0 },  /* check 42/0 */
-                    { TEST_ADD, 42, 2 },  /* check 42/2 */
-                    { DO_END }
-                  };
-                  char *buff = NULL;
-                  const char **strings = NULL;
-
-                  /* produce key #i */
-#define FMT() \
-                  char buffer[256]; \
-                  const char *name; \
-                  const long expected = (long) i * 1664525 + 1013904223; \
-                  do { \
-                    if (strings == NULL) { \
-                      snprintf(buffer, sizeof(buffer), \
-                        "http://www.example.com/website/sample/for/hashtable/" \
-                        "%ld/index.html?foo=%ld&bar", \
-                        (long) i, (long) (expected)); \
-                      name = buffer; \
-                    } else { \
-                      name = strings[i]; \
-                    } \
-                  } while(0)
-
-                  /* produce random patterns, or read from a file */
-                  if (sscanf(snum, "%lu", &count) != 1) {
-                    const off_t size = fsize(snum);
-                    FILE *fp = fopen(snum, "rb");
-                    if (fp != NULL) {
-                      buff = malloc(size);
-                      if (buff != NULL && fread(buff, 1, size, fp) == size) {
-                        size_t capa = 0;
-                        size_t i, last;
-                        for(i = 0, last = 0, count = 0 ; i < size ; i++) {
-                          if (buff[i] == 10 || buff[i] == 0) {
-                            buff[i] = '\0';
-                            if (capa == count) {
-                              if (capa == 0) {
-                                capa = 16;
-                              } else {
-                                capa <<= 1;
-                              }
-                              strings = (const char **) realloc((void*) strings, capa*sizeof(char*));
-                            }
-                            strings[count++] = &buff[last];
-                            last = i + 1;
-                          }
-                        }
-                      }
-                      fclose(fp);
-                    }
-                  }
-
-                  /* successfully read */
-                  if (count > 0) {
-                    coucal hashtable = coucal_new(0);
-                    size_t loop;
-                    for(loop = 0 ; bench[loop].type != DO_END ; loop++) {
-                      size_t i;
-                      for(i = bench[loop].offset ; i < (size_t) count
-                          ; i += bench[loop].modulus) {
-                        int result;
-                        FMT();
-                        if (bench[loop].type == DO_ADD
-                            || bench[loop].type == DO_DRY_ADD) {
-                          size_t k;
-                          result = coucal_write(hashtable, name, (uintptr_t) expected);
-                          for(k = 0 ; k < /* stash_size*2 */ 32 ; k++) {
-                            (void) coucal_write(hashtable, name, (uintptr_t) expected);
-                          }
-                          /* revert logic */
-                          if (bench[loop].type == DO_DRY_ADD) {
-                            result = result ? 0 : 1;
-                          }
-                        }
-                        else if (bench[loop].type == DO_DEL
-                            || bench[loop].type == DO_DRY_DEL) {
-                          size_t k;
-                          result = coucal_remove(hashtable, name);
-                          for(k = 0 ; k < /* stash_size*2 */ 32 ; k++) {
-                            (void) coucal_remove(hashtable, name);
-                          }
-                          /* revert logic */
-                          if (bench[loop].type == DO_DRY_DEL) {
-                            result = result ? 0 : 1;
-                          }
-                        }
-                        else if (bench[loop].type == TEST_ADD
-                            || bench[loop].type == TEST_DEL) {
-                          intptr_t value = -1;
-                          result = coucal_readptr(hashtable, name, &value);
-                          if (bench[loop].type == TEST_ADD && result
-                              && value != expected) {
-                            fprintf(stderr, "value failed for %s (expected %ld, got %ld)\n",
-                                    name, (long) expected, (long) value);
-                            exit(EXIT_FAILURE);
-                          }
-                          /* revert logic */
-                          if (bench[loop].type == TEST_DEL) {
-                            result = result ? 0 : 1;
-                          }
-                        }
-                        if (!result) {
-                          fprintf(stderr, "failed %s{%d/+%d} test on loop %ld"
-                                  " at offset %ld for %s\n",
-                                  names[bench[loop].type],
-                                  (int) bench[loop].modulus,
-                                  (int) bench[loop].offset,
-                                  (long) loop, (long) i, name);
-                          exit(EXIT_FAILURE);
-                        }
-                      }
-                    }
-                    coucal_delete(&hashtable);
-                    fprintf(stderr, "all hashtable tests were successful!\n");
-                  } else {
-                    fprintf(stderr, "Malformed number\n");
-                    exit(EXIT_FAILURE);
-                  }
-#undef FMT
-                } else {
-                  fprintf(stderr,
-                    "Option #7 needs to be followed by a number");
-                  exit(EXIT_FAILURE);
-                }
-                htsmain_free();
-                return 0;
-                break;
-              case '9': { // copy_htsopt selftest: httrack -#9
-                httrackp *from = hts_create_opt();
-                httrackp *to = hts_create_opt();
-                int err = 0;
-
-                /* from-values differ from both the to-values and the
-                   hts_create_opt() defaults (nearlink FALSE, errpage/parseall
-                   TRUE), so a copy that no-ops or just resets to defaults is
-                   caught too, not only the unsigned-guard bug. */
-                from->retry = 7; /* int field: positive control */
-                to->retry = 0;
-                from->nearlink = HTS_TRUE;
-                to->nearlink = HTS_FALSE;
-                from->errpage = HTS_FALSE;
-                to->errpage = HTS_TRUE;
-                from->parseall = HTS_FALSE;
-                to->parseall = HTS_TRUE;
-
-                copy_htsopt(from, to);
-
-                if (to->retry != 7)
-                  err = 1;
-                if (to->nearlink != HTS_TRUE)
-                  err = 1;
-                if (to->errpage != HTS_FALSE)
-                  err = 1;
-                if (to->parseall != HTS_FALSE)
-                  err = 1;
-
-                /* HTS_DEFAULT (-1) is "unspecified": copy_htsopt must skip it,
-                   leaving the target intact. Only a signed (int-backed) field
-                   can hold -1, so this also guards the type against regressing
-                   to an unsigned hts_boolean. */
-                from->parseall = HTS_DEFAULT;
-                to->parseall = HTS_TRUE;
-                copy_htsopt(from, to);
-                if (to->parseall != HTS_TRUE)
-                  err = 1;
-
-                hts_free_opt(from);
-                hts_free_opt(to);
-                printf("copy-htsopt: %s\n", err ? "FAIL" : "OK");
-                htsmain_free();
-                return err;
-              } break;
-              case 'Q': { // cookie request-header selftest: httrack -#Q
-                static t_cookie cookie;
-                char hdr[1024];
-                /* RFC 6265: bare name=value pairs, no $Version/$Path (#151). */
-                const char *expected = "Cookie: name=value; has_js=1" H_CRLF;
-                int err = 0;
-
-                const char *dom = "www.example.com";
-                int added;
-
-                cookie.max_len = (int) sizeof(cookie.data);
-                cookie.data[0] = '\0';
-                added = cookie_add(&cookie, "name", "value", dom, "/");
-                added |= cookie_add(&cookie, "has_js", "1", dom, "/");
-                /* different domain: must be filtered out */
-                added |= cookie_add(&cookie, "junk", "x", "other.org", "/");
-                if (added) {
-                  printf("cookie-header: FAIL (cookie_add setup)\n");
-                  htsmain_free();
-                  return 1;
-                }
-
-                http_cookie_header_selftest(&cookie, dom, "/", hdr,
-                                            sizeof(hdr));
-                if (strcmp(hdr, expected) != 0)
-                  err = 1;
-                if (strstr(hdr, "$Version") != NULL ||
-                    strstr(hdr, "$Path") != NULL)
-                  err = 1;
-                if (strstr(hdr, "junk") != NULL) // wrong-domain cookie leaked
-                  err = 1;
-                printf("cookie-header: %s\n", err ? "FAIL" : "OK");
-                if (err)
-                  printf("  got: %s\n", hdr);
-                htsmain_free();
-                return err;
-              } break;
               case '!':
                 HTS_PANIC_PRINTF
                   ("Option #! is disabled for security reasons");
