+++ tests/01_engine-copyopt.test
@@ -2,15 +2,16 @@
 #
 # Regression guard for the unsigned-enum sentinel trap: copy_htsopt's
 # `if (from->X > -1)` guard is always false for unsigned hts_boolean fields, so
-# they silently stop being copied. Driven by the in-process 'httrack -#9' test.
+# they silently stop being copied. Driven by the in-process 'httrack -#test=copyopt' test.
 # Keep POSIX-portable (harness runs it via $(BASH), a plain /bin/sh on macOS).
 
 set -eu
 
-# A trailing token is required; a bare '-#9' falls through to the usage screen.
-out=$(httrack -#9 run)
+# 'run' is an ignored placeholder argument.
+out=$(httrack -#test=copyopt run)
 
-# Exact-match the success line so a fall-through to usage can't pass the test.
+# Exact-match the success line so a renamed/removed test (it prints the registry
+# to stderr) can't pass.
 test "$out" = "copy-htsopt: OK" || {
     echo "expected 'copy-htsopt: OK', got: $out" >&2
     exit 1
