+++ tests/01_engine-hashtable.test
@@ -3,5 +3,7 @@
 
 set -euo pipefail
 
-# httrack internal hashtable autotest on 100K keys
-httrack -#7 100000
+# httrack internal hashtable autotest on 100K keys. Assert the success line (on
+# stderr) so a misrouted registry entry can't pass on exit code alone.
+out=$(httrack -#test=hashtable 100000 2>&1)
+printf '%s\n' "$out" | grep -q "all hashtable tests were successful!" || exit 1
