+++ tests/01_engine-mime.test
@@ -4,13 +4,13 @@
 set -euo pipefail
 
 # MIME type guessing from extension (get_httptype / give_mimext).
-# -#2 <path> prints "<path> is '<mime>'" then "and its local type is '.<ext>'".
+# -#test=mime <path> prints "<path> is '<mime>'" then "and its local type is '.<ext>'".
 
 mime() {
-    test "$(httrack -O /dev/null -#2 "$1" | head -1)" == "$1 is '$2'" || exit 1
+    test "$(httrack -O /dev/null -#test=mime "$1" | head -1)" == "$1 is '$2'" || exit 1
 }
 unknown() {
-    test "$(httrack -O /dev/null -#2 "$1" | head -1)" == "$1 is of an unknown MIME type" || exit 1
+    test "$(httrack -O /dev/null -#test=mime "$1" | head -1)" == "$1 is of an unknown MIME type" || exit 1
 }
 
 mime '/a/b.html' 'text/html'
